move SpectralPrecond into a new namespace, RandLAPACK::linops
